Remove use of templates
We prefer to copy working files and modify rather than use templates.
While this makes creating the README more difficult, it is also a lot
better.  Same for the manifest, it is more difficult but always up to
date.

The AddReadme editor and test templates were moved to be actual editors
and tests in this repo so they will not get out of date.  The editor was
also changed to not use a template.

SwitchReadme was changed to UpdateReadme and moved out to its own file.
One interesting note is that the dollar signs in the replacement text
need to be escaped with a backslash, otherwise Java String's replaceAll
will think you are trying to reference captured subsequences.
https://docs.oracle.com/javase/7/docs/api/java/util/regex/Matcher.html#replaceAll(java.lang.String)

Generator tests are currently broken so some assertions in the
NewRugArchiveProject test have been commented out.  rug 0.11.0 should
fix this and then they can be uncommented.

Fixes #10